rename inCORE service components
