+++ src/app/incore/services/coreOfferings.tsx
